ci: add script to update rockspec versions (#37)
The rockspec format requires the actual filename to contain the version
of the package - so we can't use release-please's built in
`x-release-please-version` utility.

This adds a bash script which can be run locally or in CI to accomplish
the task. Inspiration from @keelerm84's work on the iOS SDK.

It renames the rockspec & updates the `version` field contained within,
plus updates the README examples. If provided with a git username and
password, it'll commit the changes and push. Otherwise the changes are
left staged.