ci: update install-lua-sdk workflow with unified default values (#63)
The problem being solved here is that `workflow_dispatch`,
`workflow_call`, and `schedule` events are often all useful for a single
workflow.

That is - you might want a workflow to run based on:
1. A UI interaction
2. Another workflow calling it
3. On a regular schedule

The first two, and perhaps the last, are all relevant for
`install-lua-sdk`. We might have a one-off test for a particular Lua
version, triggered in Github UI (1).

We might run this after `release-please` publishes a package (2). 

And finally, we might want to ensure the package is installable every
day (3).

To make this actually work, you need to ensure default values are
assigned for each input. The cleanest way to do that seems to be some
duplication and a "ternary" technique with environment variables.