ci: fix publication workflow (#65)
This converts the `rockspec-names` action into a resuable workflow so it
can be called once and be used by multiple jobs.

Specifically, the publish workflow uses this information for:

- Invoking `luarocks publish` 
- Invoking `luarocks install` once the package is published, as a sanity
check that it succeeded

The underlying theory here is that it's better to call a reusable
workflow in a job and use that output in multiple jobs, than for each
job to call an action individually. I'm not entirely sure that's true,
but I'm getting a feel for it.