ci: remove rockspec-info.yml from release-please extra-files (#79)
This should _actually_ fix release please!

The root cause of the errors seems to have been that you cannot update a
Github workflow file using `extra-files` in a release please manifest.
This is disallowed by the Github API.

I've refactored the `rockspec-info` workflow to not need to be updated
at all. Instead, it grabs the current version from the existing
`.release-please-manifest`. I should have done this in the first place
as it's much cleaner.