docs: add EvaluationDetail table documentation (#93)
It's quite awkward to document ad-hoc tables in Lua created via the C
API. Prior to this PR, the `EvaluationDetail` table returned from
`*variationDetail` methods didn't have any explanation.

This PR adds documentation of the various fields. Since there's no
actual table to document (it's created via C API), the docs are just
sort of floating in the file as a comment.