ci: use common definition of C++ versions (#103)
This consolidates some of the places we need to specify C++ version
numbers into a single `cpp-sdk-versions.env` file.

This is a similar approach to Relay Proxy. 

Updating this file will update:
1. CI unit tests
2. Docker example tests 

But it won't update the versions found in README or in the Docker files
themselves (as they have default values for the new build arg I've
introduced.)

We need to consider the testing strategy going forward. Should we test a
min version, and then also a "current" version? And if so, is that
current going to float somehow (like using github's releases API to grab
the latest), or be pinned to the latest explicitly (via a workflow, or
Renovate or something.)